backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
python-decouple==3.8
python-dotenv==1.0.0
django-celery-beat==2.5.0
setuptools==69.0.0
psycopg2-binary==2.9.9